# Optimize Parse function